+++ Sources/Classes/Public/ImgSaveFmt.swift
@@ -20,7 +20,7 @@ import Foundation
     case MONO_4BPP = 0  ///4bpp
     case MONO_1BPP = 1  ///1bpp, transformed into 4bpp by the firmware before saving
     case MONO_4BPP_HEATSHRINK = 2 ///4bpp with Heatshrink compression, decompressed into 4bpp by the firmware before saving
-    //case MONO_4BPP_HEATSHRINK_SAVE_COMP = 3 ///4bpp with Heatshrink compression, stored compressed, decompressed into 4bpp before display
+    case MONO_4BPP_HEATSHRINK_SAVE_COMP = 3 ///4bpp with Heatshrink compression, stored compressed, decompressed into 4bpp before display
 }
 
 /// Supported image streaming format